Feature/fis 611 (#105)
* updating unit tests

* adding unit tests on station-forecast model

* added unit tests arounf ffoi station and station forcast model

* adding unit-tests

* Refactor flood test to use mocks

* Fix linting issues

* refactoring tests in flood.js

* refactored and arranged tests for floods.js

* added remaining externals

* FIS 611 refactor (#106)

* Stub out config dependency

This means the tests will run in any env regardless of env vars

* Simplifiy return value assertions

All service methods return the result from the API call unmodified so
the only test required is that this happens. The value returned is
unimportant. Changed tests to reflect this.

* Remove redundant verfy statements

All mocks will be verified by default

* Switch to object destructuring for consistency

Co-authored-by: neil mclaughlin <neil.mclaughlin@defra.gov.uk>