Change interpolation curve
With gaps in the data as a result of not plotting <0 values the cardinal
interpolation curve had interpolation artefacts in the plotted curve.
curveMonotoneX does not suffer from this as a result of preserving the
X ordering of the interpolated points.

See https://github.com/d3/d3-shape#curveMonotoneX for further details