FSR-1377 | Search bug fixes (#892)
Note: story was originally FSR-1370 hence the branch name

* Initial set of changes to allow search of english ceremonial and administrative counties

* Simplify postcode URIs

As well as for aesthetic reasons this prevents scenarios where searching by the postcode and
returned name doesn't return the same result as the postcode alone. An example of this was postcode
TQ9 7SS which was reported in feedback

* Remove unused isScotlandOrNorthernIsland flag

Also hard coded the isEngland flag as bing results are filtered to
exclude non-english results. This flag can also probably be removed next.

* Fix test name typo

* Preserve spaces in postcodes

Bing can handle most full postcodes with or without spaces between
the outcode and incode but some (?all) double digit outcodes seem
to present a problem (e.g. DE12 8EA vs DE128EA)

* Update allowed entity types and their precedence

Types and precedence reflect discussions with PO during demo

* Add slug property to search result

This is to remove code duplication and also to allow different slug creation
behaviour based on entityType

* Fix broken query and slug population

* Populate the query search box with the location name rather than the original query term
  to match the behaviour of the national page search.
* Also fixes problem with hypenated place id being used
* Fix broken URL's in station type group tabs which meant that right clicking on a group tab
  and opening in a new browser tab now works
* Added some more html helpers
* Made some of the tests more specific by testing elements rather than the whole page

* SonarCloud: remove unused parameter q

* Use Parameterised max number of results returned from bing

Note that this change depends on a corresponding change to the bing url
in the lfwconfig repo

* Add test for mix of alphanumeric and non-alphanumeric input chars

* Remove duplicate test

* Fix issue with search input not being populated consistently

* Limit results to high confidence only

* Content change for consistency

Note that we should really be using the partial search.html but that is
a bigger bit of work

* Populate search text box for postcodes with postcode only

* SonarCloud: remove unused parameter

* Add missing test for no warnings scenario

Also removed parameter 'warnings' mistakenly added to call to
headingChecker in warningBlockChecker